requirements.txt
discord.py==1.5.1
pandas==1.1.4
python-dotenv==0.15.0
